Sign up code start off
To my understandings I have completed the signup page but for some reasons the code is not working I am not sure if it is my laptop messing up or the code itself in not working. 

Can you please check what's the reason behind of the code not working @ntflix @GergoN02?